% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CDF}
\alias{CDF}
\title{Plot the CDFs of consensus clustering}
\usage{
CDF(ml, breaks = 100)
}
\arguments{
\item{ml}{list of all consensus matrices}

\item{breaks}{number of breaks}
}
\value{
a CDF plot
}
\description{
Internal function to plot the CDFs of the consensus solutions; taken
from \code{ConsensusClusterPlus}.
}
