% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cluster_tracking_plot}
\alias{cluster_tracking_plot}
\title{Generate the tracking plot}
\usage{
cluster_tracking_plot(assignment_matrix)
}
\arguments{
\item{assignment_matrix}{matrix of assigned colours (which here are equal
with cluster assignment) to every observation with one entry per specified
number of cluster (i.e. the rows are the different k and the columns the samples)}
}
\value{
tracking plot
}
\description{
Internal function to generate the tracking plot;
adapted from \code{ConsensusClusterPlus}.
}
