% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{clusterTrackingPlot}
\alias{clusterTrackingPlot}
\title{Generate the tracking plot}
\usage{
clusterTrackingPlot(m)
}
\arguments{
\item{m}{list of assigned colours to every observation with one entry per
specified number of cluster}
}
\value{
tracking plot
}
\description{
Internal function to generate the tracking plot; taken from \code{ConsensusClusterPlus}.
}
