% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{connectivity_matrix}
\alias{connectivity_matrix}
\title{Update connectivity matrix}
\usage{
connectivity_matrix(
  cluster_assignments,
  current_matrix,
  matrix_names,
  sample_key
)
}
\arguments{
\item{cluster_assignments}{vector of cluster assignment for every subject}

\item{current_matrix}{current connectivity matrix}

\item{matrix_names}{vector of the names of the rows/columns of the
connectivity matrix; usually the sorted unique names of the subjects}

\item{sample_key}{name of the subjects}
}
\value{
updated connectivity matrix
}
\description{
Internal function to update a matrix that stores information how often two
samples have gotten the same cluster assignment.
Adapted from \code{ConsensusClusterPlus}.
}
