% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_assignment}
\alias{extract_assignment}
\title{Extract one overview assignment table}
\usage{
extract_assignment(results, id_column)
}
\arguments{
\item{results}{result list where the first entry is ignored and the other
entries respond to the number of specified clusters (the kth entry means
k specified clusters); every entry needs to contain the entry
\code{consensusClass}. The cluster assignments need to have the subject
name as attribute}

\item{id_column}{character vector of the ID column in the dataset}
}
\value{
\code{data.frame} with a column named after the \code{id_column}
containing the subject name and one column with the cluster assignments for
every specified number of clusters, e.g. \code{assignment_num_clus_2}
}
\description{
This internal function generates a \code{data.frame} with the cluster
assignments for every subject across all specified numbers of clusters
}
