% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_consensus_cluster.R
\name{lcc_run}
\alias{lcc_run}
\title{Main function to run the longitudincal clustering}
\usage{
lcc_run(
  data,
  id_column,
  maxK,
  reps,
  pItem,
  model_list,
  flexmix_formula,
  verbose
)
}
\arguments{
\item{data}{a \code{data.frame} with one or several observations per subject.
It needs to contain one column that specifies to which subject the entry (row)
belongs to. This ID column is specified in \code{id_column}. Otherwise, there
are no restrictions on the column names, as the model is specified in
\code{flexmix_formula}.}

\item{id_column}{name (character vector) of the ID column in \code{data} to
identify all observations of one subject}

\item{maxK}{maximum number of clusters, default is \code{3}}

\item{reps}{number of repetitions, default is \code{10}}

\item{pItem}{fraction of samples contained in subsampled sample, default is
\code{0.8}}

\item{model_list}{either one \code{flexmix} driver or a list of \code{flexmix}
drivers of class \code{FLXMR}}

\item{flexmix_formula}{a \code{formula} object that describes the \code{flexmix}
model relative to the formula in the flexmix drivers (the dot in the flexmix
drivers is replaced, see the example). That means that you usueally only
specify the right-hand side of the formula here. However, this is not enforced
or checked to give you more flexibility over the \code{flexmix} interface}

\item{verbose}{\code{boolean} if status messages should be displayed.
Default is \code{FALSE}}
}
\value{
Returns a list with the following entries:\tabular{ll}{
\code{consensus_matrices} \tab a list of all consensus matrices where the kth entry is the consensus matrix for k specified numbers of clusters. The first entry is \code{NULL} \cr
\tab \cr
\code{found_number_clusters} \tab a list of vectors with the actual number of clusters found by \code{flexmix} where the kth entry is for k specified numbers of clusters. The first entry is \code{NULL}
}
}
\description{
Internal function to actually perform the clustering
}
