% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{my_pal}
\alias{my_pal}
\title{Generate colour palette}
\usage{
my_pal(n = 10)
}
\arguments{
\item{n}{number of colours}
}
\value{
vector of colour hex values where the first is red and the rest different
blue colours
}
\description{
Internal function to generate colour palette;
adapted from \code{ConsensusClusterPlus}.
}
