% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_functions.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Spaghetti plot for longmixr clusterings}
\usage{
plot_spaghetti(
  model,
  data,
  variable_names,
  time_variable,
  show_mean_sd_ribbon = TRUE,
  number_of_clusters = 2,
  scales = "fixed"
)
}
\arguments{
\item{model}{\code{lcc} object (output from \code{\link{longitudinal_consensus_cluster}})}

\item{data}{a \code{data.frame} that contains the variables to be plotted and
the time and ID variable used in the longmixr clustering; typically the data
used for the clustering}

\item{variable_names}{character vector of the continuous variables to be plotted}

\item{time_variable}{the name of the variable that depicts the time point of
the measurements}

\item{show_mean_sd_ribbon}{\code{boolean} if the mean and SD per variable
should be shown, the default is \code{TRUE}}

\item{number_of_clusters}{the number of clusters that should be plotted, the
default is \code{2}}

\item{scales}{\code{scales} argument of \code{facet_wrap}, the default is
\code{fixed}}
}
\value{
a \code{ggplot} object that is plotted
}
\description{
A helper function to plot spaghetti plots of continuous variables separated
by the clusters found by \code{longmixr}.
}
\details{
The spaghetti plot shows the longitudinal trajectory (defined by
\code{time_variable}) of continuous variables separated by the clusters found
by \code{\link{longitudinal_consensus_cluster}}. The provided \code{data.frame}
for \code{data} can either be the same as used in the clustering with
\code{\link{longitudinal_consensus_cluster}} or needs to contain the same
\code{id_column} as in the clustering and a \code{time_variable}.
}
\examples{
set.seed(5)
test_data <- data.frame(patient_id = rep(1:10, each = 4),
visit = rep(1:4, 10),
var_1 = c(rnorm(20, -1), rnorm(20, 3)) +
rep(seq(from = 0, to = 1.5, length.out = 4), 10),
var_2 = c(rnorm(20, 0.5, 1.5), rnorm(20, -2, 0.3)) +
rep(seq(from = 1.5, to = 0, length.out = 4), 10))
model_list <- list(flexmix::FLXMRmgcv(as.formula("var_1 ~ .")),
flexmix::FLXMRmgcv(as.formula("var_2 ~ .")))
clustering <- longitudinal_consensus_cluster(
data = test_data,
id_column = "patient_id",
max_k = 2,
reps = 3,
model_list = model_list,
flexmix_formula = as.formula("~s(visit, k = 4) | patient_id"))

plot_spaghetti(
  model = clustering,
  data = test_data,
  variable_names = "var_1",
  time_variable = "visit"
)
}
