% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ranked_bar_plot}
\alias{ranked_bar_plot}
\title{Generate item/subject consensus plot}
\usage{
ranked_bar_plot(item_consensus_matrix, cluster_colors, item_order, title)
}
\arguments{
\item{item_consensus_matrix}{matrix where the rows are different clusters,
the columns the samples and the value the item consensus of the sample with
the different clusters}

\item{cluster_colors}{colours for the clusters}

\item{item_order}{order of the items/subjects as in the solution for 2 clusters}

\item{title}{title of the plot}
}
\value{
item/subject consensus plot
}
\description{
Internal function to generate the item consensus plot;
adapted from \code{ConsensusClusterPlus}.
}
