% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rankedBarPlot}
\alias{rankedBarPlot}
\title{Generate item/subject consensus plot}
\usage{
rankedBarPlot(d, myc, cc, title)
}
\arguments{
\item{d}{data matrix}

\item{myc}{colours for the clusters}

\item{cc}{order of the items/subjects as in the solution for 2 clusters}

\item{title}{title of the plot}
}
\value{
item/subject consensus plot
}
\description{
Internal function to generate the item consensus plot; taken from \code{ConsensusClusterPlus}.
}
