% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_patients}
\alias{sample_patients}
\title{Subsample subjects}
\usage{
sample_patients(data, pSamp, id_column)
}
\arguments{
\item{data}{a \code{data.frame} with one or several observations per subject.
It needs to contain one column that specifies to which subject the entry (row)
belongs to. This ID column is specified in \code{id_column}. Otherwise, there
are no restrictions on the column names, as the model is specified in
\code{flexmix_formula}.}

\item{pSamp}{subsampling fraction, same as \code{pItem}}

\item{id_column}{name (character vector) of the ID column in \code{data} to
identify all observations of one subject}
}
\value{
List with the following entries:\tabular{ll}{
\code{subsample} \tab \code{data.frame} of the subsample \cr
\tab \cr
\code{sample_patients} \tab names from the \code{id_column} of the subsampled subjects
}
}
\description{
Internal function to subsample a fraction of \code{pSamp} subjects and
return all observations belonging to these subjects. Adapted from
\code{ConsensusClusterPlus}.
}
