% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{set_cluster_colors}
\alias{set_cluster_colors}
\title{Assign colours to cluster assignments}
\usage{
set_cluster_colors(past_ct, ct, color_names, color_list)
}
\arguments{
\item{past_ct}{cluster assignments of the previous clustering (\code{k-1} numbers of clusters)}

\item{ct}{current cluster assignments (\code{k} numbers of clusters)}

\item{color_names}{vector of colour names}

\item{color_list}{results of a previous call to \code{set_cluster_colors }}
}
\value{
List with the following entries:\tabular{ll}{
\code{1.} \tab vector of colours for every subject\cr
\tab \cr
\code{2.} \tab number; probably the number of different colours used (I haven't checked it)\cr
\tab \cr
\code{3.} \tab vector of unique colours in the first list entry
}
}
\description{
Internal function to assign colours to the cluster assignments;
adapted from \code{ConsensusClusterPlus}.
}
