% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{setClusterColors}
\alias{setClusterColors}
\title{Assign colours to cluster assignments}
\usage{
setClusterColors(past_ct, ct, colorU, colorList)
}
\arguments{
\item{past_ct}{cluster assignments of the previous clustering (\code{k-1} numbers of clusters)}

\item{ct}{current cluster assignments (\code{k} numbers of clusters)}

\item{colorU}{vector of colour names}

\item{colorList}{results of a previous call to \code{setClusterColors}}
}
\value{
List with the following entries:\tabular{ll}{
\code{1.} \tab vector of colours for every subject\cr
\tab \cr
\code{2.} \tab number; probably the number of different colours used (I haven't checked it)\cr
\tab \cr
\code{3.} \tab vector of unique colours in the first list entry
}
}
\description{
Internal function to assign colours to the cluster assignments;
taken from \code{ConsensusClusterPlus}.
}
