% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{triangle}
\alias{triangle}
\title{Generate triangle matrix}
\usage{
triangle(m, mode = 1)
}
\arguments{
\item{m}{matrix}

\item{mode}{flag which matrix should be returned; can be \code{1, 2} or \code{3}}
}
\value{
matrix; if \code{mode = 1} return the lower triangle as vector;
if \code{mode = 2} return the transformed upper triangle (so that it is the lower one now)
as a matrix with the rest of the entries as \code{NA}; if \code{mode = 3}
return a matrix where the lower triangle is replaced by the upper triangle
}
\description{
Internal function to generate a triangle matrix; taken from \code{ConsensusClusterPlus}.
}
